Moved files and implemented Messages
Moved files around to follow Page>Components structure.
Refactored import paths to use absolute paths.
Reduced comments to only key comments.
Implemented Messages & Error Messages.
Added a ',' to last item in routedict so that future additions to that object would not show alterations to previous line of code.
